Update README.md
I had to add a few lines to _config.yml in order to get this to work from a fresh install of jekyll. I think this has something to do with paginate being removed from the default installation.

Thanks! Super cool theme and great to get to use Semantic UI with Jekyll.